Rename Profile fles to GenProfile
